WIP: Templates refactor POC
