+++ docs/content/templates/home/index.mdx
@@ -5,7 +5,7 @@ description: Provide a strong first impression.
 version: 1.0.0
 previewPath: /
 figmaTemplateNodeId: 8410%3A83756
-githubTemplatePath: /example-site/pages/index.tsx
+githubTemplatePath: /docs/content/templates/home/Home.tsx
 storybookPath: /story/templates-home-page--home
 ---
